requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0

opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
